requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==4.1.6
django-bootstrap-v5==1.0.11
django-filter==22.1
Pillow==9.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
